+++ projects/hipcub/benchmark/benchmark_block_radix_sort.cpp
@@ -285,12 +285,12 @@ void add_benchmarks(primbench::executor& executor)
     BENCHMARK_TYPE(int8_t, 320);
     BENCHMARK_TYPE(int8_t, 512);
 
-    BENCHMARK_TYPE(long long, 64);
-    BENCHMARK_TYPE(long long, 128);
-    BENCHMARK_TYPE(long long, 192);
-    BENCHMARK_TYPE(long long, 256);
-    BENCHMARK_TYPE(long long, 320);
-    BENCHMARK_TYPE(long long, 512);
+    BENCHMARK_TYPE(int64_t, 64);
+    BENCHMARK_TYPE(int64_t, 128);
+    BENCHMARK_TYPE(int64_t, 192);
+    BENCHMARK_TYPE(int64_t, 256);
+    BENCHMARK_TYPE(int64_t, 320);
+    BENCHMARK_TYPE(int64_t, 512);
 
     BENCHMARK_TYPE(custom_int_t, 64);
     BENCHMARK_TYPE(custom_int_t, 128);
